Moved multiqc() into multiqc/multiqc.py for more multiqc
